added new presenters
